Add reading of first half of file

+++ Project one/src/Edge.java
@@ -4,10 +4,10 @@ public class Edge {
 	public Node firstNode;
 	public Node secondNode;
 	
-	public int edgeWeight;
+	public float edgeWeight;
 	
 
-	public Edge(Node firstNode, Node secondNode, int edgeWeight) {
+	public Edge(Node firstNode, Node secondNode, float edgeWeight) {
 		this.firstNode = firstNode;
 		this.secondNode = secondNode;
 		this.edgeWeight = edgeWeight;
